Change config to allow class name or evaluator string

+++ src/factories/digi/SiliconPulseGeneration_factory.h
@@ -23,9 +23,9 @@ private:
 
   PodioOutput<edm4hep::TimeSeries> m_out_pulses{this};
 
-  ParameterRef<double> m_timestep{this, "timestep", config().timestep};
-  ParameterRef<double> m_ignore_thres{this, "ignoreThreshold", config().ignore_thres};
-  ParameterRef<double> m_min_sampling_time{this, "minSamplingTime", config().min_sampling_time};
+  ParameterRef<float> m_timestep{this, "timestep", config().timestep};
+  ParameterRef<float> m_ignore_thres{this, "ignoreThreshold", config().ignore_thres};
+  ParameterRef<float> m_min_sampling_time{this, "minSamplingTime", config().min_sampling_time};
   ParameterRef<int> m_max_time_bins{this, "maxTimeBins", config().max_time_bins};
 
   Service<AlgorithmsInit_service> m_algorithmsInit{this};
